server/requirements.txt
channels==3.0.4
channels-redis==3.3.1
Django==4.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
Pillow==8.4.0
mysqlclient
pytest-asyncio==0.16.0
pytest-django==4.5.2
